More crash fixed when there are 0 songs on the device.
There was a crash when tapping the shuffle button on the gallery screen, and the shuffle button at the top of the gallery list view screen.

Fixes - First, we should not attempt to create an MPMediaItemCollection with zero or nil items. Second, we should not attempt to set the music players queue with a nil collection. Third, we should make sure an index path exists before attempting to reload using that index path on the gallery view controller.